feat: track enemy kills feat: move weapon scroll logic to Singleton
