+++ game/core/src/main/de/dhbw/tinf22b6/gameobject/Enemy.java
@@ -23,6 +23,7 @@ import de.dhbw.tinf22b6.ai.Box2DLocation;
 import de.dhbw.tinf22b6.ai.EnemySteeringBehaviour;
 import de.dhbw.tinf22b6.util.Constants;
 import de.dhbw.tinf22b6.util.EntitySystem;
+import de.dhbw.tinf22b6.util.PlayerStatistics;
 import de.dhbw.tinf22b6.util.SteeringUtils;
 import de.dhbw.tinf22b6.weapon.EnemyWeapon;
 import de.dhbw.tinf22b6.weapon.Weapon;
@@ -141,6 +142,7 @@ public abstract class Enemy extends MobGameObject implements Steerable<Vector2> 
         this.health--;
         if (health == 0) {
             this.remove = true;
+            PlayerStatistics.instance.enemyKilled();
         }
         Gdx.audio.newSound(Gdx.files.internal("sfx/hitSound.mp3")).play(1);
     }
