+++ game/core/src/main/de/dhbw/tinf22b6/util/PlayerStatistics.java
@@ -1,6 +1,8 @@
 package de.dhbw.tinf22b6.util;
 
 import com.badlogic.gdx.Gdx;
+import de.dhbw.tinf22b6.weapon.Ak;
+import de.dhbw.tinf22b6.weapon.M4;
 import de.dhbw.tinf22b6.weapon.Weapon;
 import java.util.ArrayList;
 import java.util.List;
@@ -13,12 +15,15 @@ public class PlayerStatistics {
     private int enemies_kills;
     private float gameTime;
     private List<Weapon> weapons;
+    private int currentWeaponIndex;
 
     // singleton: prevent instantiation from other classes
     private PlayerStatistics() {}
 
     public void init() {
         this.weapons = new ArrayList<>();
+        weapons.add(new Ak());
+        this.canSwitchWeapon = true;
         this.hp = 5;
         this.coins = 0;
         this.enemies_kills = 0;
@@ -49,7 +54,7 @@ public class PlayerStatistics {
         return enemies_kills;
     }
 
-    public void setEnemies_kills() {
+    public void enemyKilled() {
         this.enemies_kills++;
     }
 
@@ -67,5 +72,37 @@ public class PlayerStatistics {
 
     public List<Weapon> getWeapons() {
         return weapons;
+    }
+
+    public Weapon getCurrentWeapon() {
+        return weapons.get(currentWeaponIndex);
+    }
+
+    private boolean canSwitchWeapon;
+
+    public void cycleWeapon(boolean direction) {
+        List<Weapon> inventory = PlayerStatistics.instance.getWeapons();
+        if (canSwitchWeapon && inventory.size() > 1) {
+            canSwitchWeapon = false;
+            currentWeaponIndex =
+                    Math.floorMod(direction ? currentWeaponIndex - 1 : currentWeaponIndex + 1, inventory.size());
+            Gdx.audio
+                    .newSound(Gdx.files.internal("sfx/change_weapon.mp3"))
+                    .play(Gdx.app.getPreferences("Controls").getFloat("sfx"));
+        }
+        new Thread(() -> {
+                    try {
+                        Thread.sleep(500);
+                        canSwitchWeapon = true;
+                    } catch (InterruptedException e) {
+                        throw new RuntimeException(e);
+                    }
+                })
+                .start();
+    }
+
+    public void pickupWeapon() {
+        this.weapons.add(new M4());
+        this.currentWeaponIndex = weapons.size() - 1;
     }
 }
